+++ src/core/component/watch/bind.ts
@@ -7,7 +7,6 @@
  */
 
 import { getPropertyInfo } from 'core/component/reflection';
-import type { EventEmitterLike } from 'core/async';
 
 import { beforeHooks } from 'core/component/const';
 import { customWatcherRgxp } from 'core/component/watch/const';
@@ -144,7 +143,7 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 				};
 
 				let
-					handler;
+					handler: AnyFunction;
 
 				// Right now we need to create a wrapper for our "raw" handler,
 				// because there are some conditions for the watcher:
@@ -258,7 +257,7 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 
 							} else {
 								const watch = (watcherCtx) =>
-									$a.on(<EventEmitterLike>watcherCtx, watchPath, handler, eventParams, ...watchInfo.args ?? []);
+									$a.on(watcherCtx, watchPath, <AnyFunction>handler, eventParams, ...watchInfo.args ?? []);
 
 								if (Object.isPromise(watcherCtx)) {
 									$a.promise(watcherCtx, asyncParams).then(watch, stderr);
@@ -271,12 +270,28 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 							return;
 						}
 
-						const
-							unwatch = $watch.call(component, p.info ?? getPropertyInfo(watchPath, component), watchInfo, handler);
+						// eslint-disable-next-line prefer-const
+						let link, unwatch;
 
-						if (Object.isFunction(unwatch)) {
-							$a.worker(unwatch, asyncParams);
-						}
+						const emitter = (_, wrappedHandler) => {
+							handler = wrappedHandler;
+
+							$a.worker(() => {
+								if (link != null) {
+									$a.off(link);
+								}
+							}, asyncParams);
+
+							return () => unwatch?.();
+						};
+
+						link = $a.on(emitter, 'mutation', handler, {
+							...asyncParams,
+							group: `${asyncParams.group ?? ''}:suspend`
+						});
+
+						const toWatch = p.info ?? getPropertyInfo(watchPath, component);
+						unwatch = $watch.call(component, toWatch, watchInfo, handler);
 					}, stderr);
 
 				} else {
@@ -292,26 +307,42 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 							unsafe.$on(watchPath, handler);
 
 						} else {
-							const watch = (watcherCtx) =>
-								$a.on(<EventEmitterLike>watcherCtx, watchPath, handler, eventParams, ...watchInfo.args ?? []);
+							const addListener = (watcherCtx) =>
+								$a.on(watcherCtx, watchPath, handler, eventParams, ...watchInfo.args ?? []);
 
 							if (Object.isPromise(watcherCtx)) {
-								$a.promise(watcherCtx, asyncParams).then(watch, stderr);
+								$a.promise(watcherCtx, asyncParams).then(addListener, stderr);
 
 							} else {
-								watch(watcherCtx);
+								addListener(watcherCtx);
 							}
 						}
 
 						continue;
 					}
 
-					const
-						unwatch = $watch.call(component, p.info ?? getPropertyInfo(watchPath, component), watchInfo, handler);
+					// eslint-disable-next-line prefer-const
+					let link, unwatch;
 
-					if (Object.isFunction(unwatch)) {
-						$a.worker(unwatch, asyncParams);
-					}
+					const emitter = (_, wrappedHandler) => {
+						handler = wrappedHandler;
+
+						$a.worker(() => {
+							if (link != null) {
+								$a.off(link);
+							}
+						}, asyncParams);
+
+						return () => unwatch?.();
+					};
+
+					link = $a.on(emitter, 'mutation', handler, {
+						...asyncParams,
+						group: `${asyncParams.group ?? ''}:suspend`
+					});
+
+					const toWatch = p.info ?? getPropertyInfo(watchPath, component);
+					unwatch = $watch.call(component, toWatch, watchInfo, handler);
 				}
 			}
 		};